Implemented server handler
