+++ intmodules/__init__.py
@@ -2,4 +2,4 @@ from intmodules.log_module import config_logger
 from intmodules.log_module import logger
 from intmodules.preprocessor import preprocess_data
 from intmodules.json_validation import validate_request
-from intmodules.mpb_handler import MpbHandler
+from intmodules.server_handler import MPBRequestHandler
