+++ intmodules/log_module.py
@@ -3,7 +3,7 @@ import logging as log
 import logging.config
 
 
-log_path = "../services/logs\\"
+log_path = "..\\logs\\"
 logger = logging.getLogger()
